RTOS

+++ Test_Application/BSW/Src/HCSR04.c
@@ -23,7 +23,7 @@ uint32_t IC_Val1 = 0;
 uint32_t IC_Val2 = 0;
 uint32_t Difference = 0;
 uint8_t Is_First_Captured = 0;
-uint8_t Distance = 0;
+uint16_t Distance = 0;
 
 #define TRIG_PIN GPIO_PIN_9
 #define TRIG_Port GPIOA
